refactor(desktop): extract extension
