+++ core/common/src/desktopMain/kotlin/com/neoutils/neoregex/core/common/util/DragHandler.kt
