+++ core/design-system/src/commonMain/kotlin/com/neoutils/neoregex/core/designsystem/theme/Dimensions.kt
